refactor - Place logic & view SRP
